ocr_backend/requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
pillow
